feat: disable delete for posts that were not created by user

+++ section13/App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -26,7 +28,7 @@ class ListingController
   {
 
     $listings = $this->db->query('
-      SELECT * FROM workopia.listings;
+      SELECT * FROM workopia.listings ORDER BY created_at DESC;
     ')->fetchAll();
 
     loadView('listings/index', ['listings' => $listings]);
@@ -76,7 +78,7 @@ class ListingController
 
     $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
 
     // can pass in the function name and arguments to array map, running the function on for each array item
     $newListingData = array_map('sanitize', $newListingData);
@@ -146,6 +148,12 @@ class ListingController
     if (!$listing) {
       ErrorController::notFound('Listing not found');
       return;
+    }
+
+    if (!Authorization::isOwner($listing['user_id'])) {
+      $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+
+      return redirect('/listings/' . $listing['id']);
     }
 
     $this->db->query('DELETE FROM listings WHERE id = :id', $params);
